:bug: Fix: place lat, lng double 컬럼으로 변경

+++ backend/src/main/java/com/ssapin/backend/api/domain/dto/response/PlaceMapResponse.java
@@ -64,8 +64,8 @@ public class PlaceMapResponse {
         private long placeId;
         private long itemId;
         private String title;
-        private float lat;
-        private float lng;
+        private double lat;
+        private double lng;
         private String address;
 
         public PlaceResponse(Place place) {
